requirements.txt
rdflib
requests
SPARQLWrapper
pyyaml
sslyze
selenium
